+++ RELEASE_NOTES.md
@@ -400,3 +400,4 @@
 - Added function `iot_data_alloc_const_f64` to allocate a const double data value
 - Fixed reading of values which exceed the size of int64_t and uint64_t in `iot_data_primitive_from_json`
 - Support added for Ubuntu 24.04
+- Support added for Fedora 40 and removed for Fedora 36